Diagram from which SQL for creating database was generated.
Inside pgerd file is json, so it can be processed by some script
to output as graphviz diagram that can be rendered in Chen's
notation.

Even without changing notation it would be nice to output to pdf,
rather than into pdf image.

I want to save this file to reuse in the future, because now
i will cut parts of the diagram to create smaller diagrams, that can
be presented in coursework paper.